Refactor payload logic and add buildx support for faster builds
