+++ llm/llm_linux.go
@@ -1,11 +1,7 @@
 package llm
 
 import (
-	"embed"
 	"syscall"
 )
-
-//go:embed build/linux/*/*/bin/*
-var libEmbed embed.FS
 
 var LlamaServerSysProcAttr = &syscall.SysProcAttr{}
