+++ llm/payloads_unused.go
@@ -1,0 +1,8 @@
+//go:build !linux && !darwin
+
+package llm
+
+import "embed"
+
+// unused on windows
+var libEmbed embed.FS
